fix: 13 defectos reales de producción detectados en la revisión
Escrituras perdidas en silencio:
- BaseCrudOperations: los métodos de escritura por defecto devolvían null/0
  en vez de fallar; un repositorio que olvidara implementarlos borraba datos
  sin avisar. Ahora lanzan UnsupportedOperationException.

Proxies de Spring saltados por autoinvocación (this.metodo()):
- WebhookDispatcherService.retryDueDeliveries: el reintento corría dentro de
  la transacción del planificador, así que un suscriptor lento retrasaba a
  todos los demás vencidos. Ahora va por self.attempt().

Pagos que se perdían en silencio:
- Los webhooks devolvían 200 aunque el procesado fallara, y la pasarela no
  reintentaba nunca. Nueva WebhookProcessingException.
- El webhook de tienda creaba pedidos sin dirección de envío.

Datos de otro cliente:
- MeAffiliateController: cualquier afiliado podía desactivar el código de
  referido de otro (IDOR). setCodeActive comprueba la propiedad.

Cobros mal calculados:
- CustomerSubscriptionUseCase: el checkout de Stripe iba con
  user@example.com y URLs de localhost; y el barrido de pruebas gratuitas
  podía cancelar suscripciones DE PAGO sin código de plan.
- SourcingUseCase: la cuota salía siempre del plan FREE, así que quien
  pagaba recibía 5 búsquedas en vez de 30/60.

Otros:
- EmailQueueService: un fallo de envío marcaba FAILED a la primera; ahora
  quedan en cola hasta 5 intentos.
- La ficha técnica guardaba el idioma en mayúsculas y la consulta compara
  con '=', así que esas especificaciones no se mostraban.

+++ src/main/java/com/nexaplatform/dropshipping/api/mapper/CatalogStorefrontReadService.java
@@ -289,7 +289,9 @@ public class CatalogStorefrontReadService {
                 .collect(Collectors.toMap(ProductEntity::getId, p -> p, (a, b) -> a));
         List<ProductSummaryView> all = productIds.stream().map(byId::get).filter(Objects::nonNull)
                 .map(p -> productMapper.toSummary(p, lang)).toList();
-        int from = Math.min(page * safe, all.size());
+        // (long) para que un ?page enorme no desborde el int y deje un índice negativo que revienta
+        // el subList con un 500.
+        int from = (int) Math.min((long) page * safe, all.size());
         int to = Math.min(from + safe, all.size());
         return PageResponse.from(new PageImpl<>(all.subList(from, to), pageable, all.size()));
     }
